Gjør heller service conditional, enn controlleren
